requirements.txt
Flask==3.0.2
gunicorn==21.2.0
gspread==5.12.0
pandas==2.1.4
gspread-dataframe==3.3.1
google-auth==2.40.3
google-api-python-client==2.122.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
httplib2==0.22.0
oauth2client==4.1.3
